Fix home directory under cross compilation.
Our use of os/user.Current() is legit, but it only works when you
compile the app natively. If we wanted to distribute for Windows or OSX,
we'd have to compile on those platforms. That would be a total chore
when we have a perfectly good CI service to do this for us, but only
under Linux. As far as I know this is the only issue we'll have, so I'm
working around it. This won't be fixed in Go because they don't consider
it a bug if you follow the link in the comment in this commit.